rename aa-table-vpromms to road-table
